2_001 Diseño Pantalla Principal  - Diseño de pantalla principal  - Autenticacion exitosa me lleva a la pantalla principal  - Boton de logout en pantalla principal
