+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,8 @@
         android:supportsRtl="true"
         android:theme="@style/HuaweiTheme"
         tools:ignore="AllowBackup">
-        <activity android:name=".AuthActivity">
+        <activity android:name=".main.MainActivity"/>
+        <activity android:name=".auth.AuthActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
