+++ app/src/main/res/drawable/color_btn.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="#EF484B" />
+    <corners android:radius="20dp" />
+
+</shape>
