+++ app/src/main/res/drawable/color_dist.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="#0A000000" />
+    <corners android:radius="20dp" />
+
+</shape>
