+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,9 @@
 <resources>
-    <string name="app_name">SelfieCam</string>
+    <string name="app_name">Selfie Cam</string>
     <string name="auth_title">Bienvenido a HSelfie Cam</string>
     <string name="auth_desc">Bienvenido a la mejor camara de selfies. Para usarme, por favor ingresa con tu Huawei ID. Te tomara menos de 1 minuto ingresar a probar esta maravillosa aplicación</string>
+    <string name="capture_desc">Existen dos formas de capturar tu selfie. La camara toma tu foto si hay varias personas sonriendo. O toma la foto si hay una persona mas cerca sonriendo. Tu decides!</string>
+    <string name="capture_title">Captura tu sonrisa!</string>
+    <string name="most_people">La Mayoria Sonrie!</string>
+    <string name="nearest_people">La persona mas cercana Sonrie!</string>
 </resources>
